# openapi.yaml
openapi: 3.1.1
info:
  version: "1.0"
  title: AppFirst
servers:
  - url: http://localhost:3000
components:
  schemas:
    helloResponse:
      type: object
    Product:
      type: object
      required:
        - name
        - price
        - category
      properties:
        name: 
          type: string
          minLength: 2
          maxLength: 40
        description:
          type: string
          maxLength: 500
        price: 
          type: number
          minimum: 0
          multipleOf: 0.01
        category:
          type: array
          enum: [electronics , books , clothes , food]
        tags: 
          type: array
          items: 
            type: string
          minItems: 1
        inStock:
          type: boolean
          additionalProperties:
            type: string
        ratings:
            type: array
            items:
              type: object
              required:
              - score
              - comment
              properties:
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 200
    ProductCreate:
      type: object
      required:
        - name
        - price
        - category
      properties:
        name: 
          type: string
          minLength: 2
          maxLength: 40
        description:
          type: string
          maxLength: 500
        price: 
          type: number
          minimum: 0
          multipleOf: 0.01
        category:
          type: array
          enum: [electronics , books , clothes , food]
        tags: 
          type: array
          items: 
            type: string
          minItems: 1
        inStock:
          type: boolean
paths:
  /hello:
    get:
      summary: hello world
      responses:
        '200':
          description: Respuesta completa
          content:
            application/json:
              schema:
                type: object

  /users:
    post:
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - age
              - email
              properties:
                name:
                  type: string
                  minLength: 2
                age:
                  type: integer
                  minimum: 18
                email:
                  type: string
                  format: email
      responses:
        201:
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  age:
                    type: integer
                  email:
                    type: string
  /users/{id}:
    get:
      description: "Obtener un usuario por el ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: usuario encontrado 
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name: 
                    type: string
        '404':
            description: usuario no encontrado 
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string  
    post:
      summary:  actualizar un usuario por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              required:
                - name
                - age
                - email
              properties:
                name:
                  type: string
                  minLength: 2
                email:
                  type: string
                  format: email
                age:
                  type: integer
                  minimum: 18

          

      responses:
        '200':
          description: usuario actualizado 
          content:
            application/json:
              schema: 
                type: object
                properties:
                  id:
                    type: integer
                  name: 
                    type: string
                  age:
                    type: integer
                  email:
                    type: string
        '404':
          description: usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string 
  /productos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Obtener un producto por ID
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      summary: Actualizar un producto por ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Eliminar un producto por ID
      responses:
        '204':
          description: Producto eliminado


            